server/requirements/base.txt
Django==4.0.2
django-environ==0.8.1
shortuuid==1.0.8

djangorestframework==3.13.1
